refactor: Modify structure to follow conventional architecture
Comes with the addition of an application c and h file that initializes the necessary controllers and moves code that was originally in main into that layer.
