+++ lib/Application/application.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void application_init();
+void application_run();
